requirements.txt
numpy~=1.20.1
pillow~=8.2.0
matplotlib~=3.3.4